[TASK] Move page tree logic into new helper Typo3PageTree
