+++ packages/screenshots/Classes/Runner/Codeception/Support/Helper/Typo3Navigation.php
@@ -14,22 +14,11 @@ namespace TYPO3\CMS\Screenshots\Runner\Codeception\Support\Helper;
 
 use Codeception\Module;
 use Codeception\Module\WebDriver;
-use Facebook\WebDriver\Exception\ElementNotInteractableException;
-use Facebook\WebDriver\Exception\NoSuchElementException;
-use Facebook\WebDriver\Remote\RemoteWebElement;
-use Facebook\WebDriver\WebDriverBy;
 use Facebook\WebDriver\WebDriverElement;
 use TYPO3\TestingFramework\Core\Acceptance\Helper\Login;
 
 /**
  * Helper to support comfortable navigation of the TYPO3 backend.
- *
- * This helper contains a slightly adapted copy of class AbstractPageTree of the typo3/testing-framework package.
- * It should be integrated there ideally. Currently it differs by:
- * - adding the actions to the actor instead of providing an own class
- *   to prevent additional injections in testing classes
- * - considering the configuration param "wait" of module "WebDriver"
- *   when opening the page tree path and the page tree is not visible immediately
  *
  * This helper contains a slightly adapted copy of trait FrameSteps of the typo3/testing-framework package.
  * It should be integrated there ideally. Currently it differs by:
@@ -46,10 +35,6 @@ use TYPO3\TestingFramework\Core\Acceptance\Helper\Login;
 class Typo3Navigation extends Module
 {
     // Selectors
-    protected string $pageTreeFrameSelector = '#typo3-pagetree';
-    protected string $pageTreeSelector = '#typo3-pagetree-treeContainer';
-    protected string $treeItemSelector = 'g.nodes > .node';
-    protected string $treeItemAnchorSelector = 'text.node-name';
     protected string $openedModalSelector = '.modal.show';
     protected string $openedModalButtonContainerSelector = '.modal.show .modal-footer';
 
@@ -146,7 +131,7 @@ class Typo3Navigation extends Module
     {
         $frameSelector = ['class' => 'scaffold-modulemenu'];
         $offsetY = $offsetY + $this->_getHeaderHeight() + 5;
-        $this->scrollFrameTo($frameSelector, $toSelector, $offsetX, $offsetY);
+        $this->_scrollFrameTo($frameSelector, $toSelector, $offsetX, $offsetY);
     }
 
     public function _getHeaderHeight(): int
@@ -175,7 +160,7 @@ class Typo3Navigation extends Module
     public function scrollModuleMenuToTop(): void
     {
         $frameSelector = ['class' => 'scaffold-modulemenu'];
-        $this->scrollFrameToTop($frameSelector);
+        $this->_scrollFrameToTop($frameSelector);
     }
 
     /**
@@ -184,69 +169,7 @@ class Typo3Navigation extends Module
     public function scrollModuleMenuToBottom(): void
     {
         $frameSelector = ['class' => 'scaffold-modulemenu'];
-        $this->scrollFrameToBottom($frameSelector);
-    }
-
-    /**
-     * Scroll the page tree up to show the given element at the top.
-     *
-     * Make sure that the element is visible in the browser window before using this action. Otherwise, the browser
-     * cannot determine the position of the element because the page tree is an SVG element and the browser has limited
-     * capabilities with respect to SVGs.
-     *
-     * ``` php
-     * <?php
-     * $I->scrollPageTreeTo('.node.identifier-0_32');
-     * ?>
-     * ```
-     *
-     * @param string $toSelector
-     * @param int $offsetX
-     * @param int $offsetY
-     */
-    public function scrollPageTreeTo(string $toSelector, int $offsetX = 0, int $offsetY = 0): void
-    {
-        $frameSelector = ['id' => 'typo3-pagetree-tree'];
-        $offsetY = $offsetY + $this->_getPageTreeToolbarHeight() + $this->_getHeaderHeight() + 3;
-        $this->scrollFrameTo($frameSelector, $toSelector, $offsetX, $offsetY);
-    }
-
-    public function _getPageTreeToolbarHeight(): int
-    {
-        /** @var WebDriverElement[] $elements */
-        $elements = $this->getWebDriver()->_findElements(['id' => 'typo3-pagetree-toolbar']);
-        if (count($elements) > 0) {
-            return $elements[0]->getSize()->getHeight();
-        }
-        return 0;
-    }
-
-    public function _getPageTreeScrollHeight(): int
-    {
-        /** @var WebDriverElement[] $elements */
-        $elements = $this->getWebDriver()->_findElements(['id' => 'typo3-pagetree-tree']);
-        if (count($elements) > 0) {
-            return $this->getWebDriver()->executeJS("return arguments[0].scrollHeight", $elements);
-        }
-        return 0;
-    }
-
-    /**
-     * Move the page tree of the main frame to top.
-     */
-    public function scrollPageTreeToTop(): void
-    {
-        $frameSelector = ['id' => 'typo3-pagetree-tree'];
-        $this->scrollFrameToTop($frameSelector);
-    }
-
-    /**
-     * Move the page tree of the main frame to the bottom.
-     */
-    public function scrollPageTreeToBottom(): void
-    {
-        $frameSelector = ['id' => 'typo3-pagetree-tree'];
-        $this->scrollFrameToBottom($frameSelector);
+        $this->_scrollFrameToBottom($frameSelector);
     }
 
     /**
@@ -266,7 +189,7 @@ class Typo3Navigation extends Module
     {
         $frameSelector = ['class' => 'module'];
         $offsetY = $offsetY + $this->_getModuleHeaderHeight() + 10;
-        $this->scrollFrameTo($frameSelector, $toSelector, $offsetX, $offsetY);
+        $this->_scrollFrameTo($frameSelector, $toSelector, $offsetX, $offsetY);
     }
 
     public function _getModuleHeaderHeight(): int
@@ -295,7 +218,7 @@ class Typo3Navigation extends Module
     public function scrollModuleBodyToTop(): void
     {
         $frameSelector = ['class' => 'module'];
-        $this->scrollFrameToTop($frameSelector);
+        $this->_scrollFrameToTop($frameSelector);
     }
 
     /**
@@ -304,7 +227,7 @@ class Typo3Navigation extends Module
     public function scrollModuleBodyToBottom(): void
     {
         $frameSelector = ['class' => 'module'];
-        $this->scrollFrameToBottom($frameSelector);
+        $this->_scrollFrameToBottom($frameSelector);
     }
 
     /**
@@ -321,7 +244,7 @@ class Typo3Navigation extends Module
      *
      * @see \Codeception\Module\WebDriver::scrollTo
      */
-    protected function scrollFrameTo($frameSelector, string $toSelector, int $offsetX = 0, int $offsetY = 0): void
+    public function _scrollFrameTo($frameSelector, string $toSelector, int $offsetX = 0, int $offsetY = 0): void
     {
         $webDriver = $this->getWebDriver();
 
@@ -346,7 +269,7 @@ class Typo3Navigation extends Module
      * @param string|array $frameSelector
      * @return void
      */
-    protected function scrollFrameToTop($frameSelector): void
+    public function _scrollFrameToTop($frameSelector): void
     {
         $frameElement = $this->getWebDriver()->_findElements($frameSelector)[0];
         $this->getWebDriver()->executeJS("arguments[0].scrollTop = 0", [$frameElement]);
@@ -358,62 +281,10 @@ class Typo3Navigation extends Module
      * @param string|array $frameSelector
      * @return void
      */
-    protected function scrollFrameToBottom($frameSelector): void
+    public function _scrollFrameToBottom($frameSelector): void
     {
         $frameElement = $this->getWebDriver()->_findElements($frameSelector)[0];
         $this->getWebDriver()->executeJS("arguments[0].scrollTop = arguments[0].scrollHeight", [$frameElement]);
-    }
-
-    /**
-     * Open the given hierarchical path in the page tree and click the last page.
-     *
-     * Make sure that the pages of the hierarchical path are visible in the browser window before using this action.
-     * Otherwise, the browser cannot find all page nodes of the path because the page tree is an SVG element and
-     * the browser has limited capabilities with respect to SVGs.
-     *
-     * Example to open "styleguide -> elements basic" page:
-     * [
-     *    'styleguide TCA demo',
-     *    'elements basic',
-     * ]
-     *
-     * @param string[] $path
-     * @throws \Codeception\Exception\ModuleException
-     */
-    public function openPageTreePath(array $path): void
-    {
-        $this->switchToMainFrame();
-
-        $webDriver = $this->getWebDriver();
-        $webDriver->seeElement(['css' => $this->pageTreeSelector]);
-        $pageTree = $webDriver->_findElements(['css' => $this->pageTreeSelector])[0];
-        foreach ($path as $pageName) {
-            $pageTree = $this->ensureTreeNodeIsOpen($pageName, $pageTree);
-        }
-        $pageTree->findElement(WebDriverBy::cssSelector($this->treeItemAnchorSelector))->click();
-    }
-
-    /**
-     * Search for a page node in the provided page tree and open it.
-     *
-     * @param string $nodeText
-     * @param RemoteWebElement $pageTree
-     * @return RemoteWebElement
-     * @throws \Codeception\Exception\ModuleException
-     */
-    protected function ensureTreeNodeIsOpen(string $nodeText, RemoteWebElement $pageTree): RemoteWebElement
-    {
-        $webDriver = $this->getWebDriver();
-        $webDriver->see($nodeText, $this->treeItemSelector);
-        $node = $pageTree->findElement(WebDriverBy::xpath('//*[text()=\'' . $nodeText . '\']/..'));
-        try {
-            $node->findElement(WebDriverBy::cssSelector('.chevron.collapsed'))->click();
-        } catch (NoSuchElementException $e) {
-            // element not found so it may be already opened...
-        } catch (ElementNotInteractableException $e) {
-            // another possible exception if the chevron isn't there ... depends on facebook driver version
-        }
-        return $node;
     }
 
     public function waitForModalDialogInMainFrame(): void
